THER-49 added basic service layers for Appointments, diagnosis, treatment

+++ src/main/java/org/dci/theratrack/repository/TherapistRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TherapistRepository extends JpaRepository<Therapist, Integer> {
+public interface TherapistRepository extends JpaRepository<Therapist, Long> {
 }
